computer_science/deep_learning/code/neural-magic-eye/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
opencv-python>=4.5.0
matplotlib>=3.5.0
tqdm>=4.60.0
Pillow>=9.0.0
scikit-image>=0.19.0
transformers>=4.30.0 